test: split up test file into two separate files
